src/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xpressai"
version = "0.1.0"
description = "Agent runtime - the Phusion Passenger for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Xpress AI", email = "hello@xpress.ai" }
]
keywords = ["ai", "agents", "llm", "runtime", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "click>=8.0",
    
    # Database
    "sqlite-vec>=0.1.0",
    
    # Agent SDK
    "claude-agent-sdk>=0.1.0",
    
    # Docker
    "docker>=7.0",
    
    # Web UI
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
    
    # TUI
    "textual>=0.50",
    "rich>=13.0",
    
    # Scheduling
    "apscheduler>=3.10",
    
    # Embeddings (local)
    "sentence-transformers>=2.5",
    
    # Local inference (optional, but useful)
    "huggingface-hub>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

llama = [
    "llama-cpp-python>=0.2",
]

[project.scripts]
xpressai = "xpressai.cli.main:cli"

[project.urls]
Homepage = "https://github.com/xpressai/xpressai"
Documentation = "https://docs.xpress.ai"
Repository = "https://github.com/xpressai/xpressai"

[tool.hatch.build.targets.wheel]
packages = ["src/xpressai"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=xpressai --cov-report=term-missing"
